Factor tt::is_a into its own file
