+++ src/DataStructures/DataBox/DataBoxTag.hpp
@@ -22,6 +22,7 @@
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TypeTraits.hpp"
 #include "Utilities/TypeTraits/CreateIsCallable.hpp"
+#include "Utilities/TypeTraits/IsA.hpp"
 
 namespace Tags {
 /*!
