+++ src/DataStructures/Variables.hpp
@@ -31,6 +31,7 @@
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TaggedTuple.hpp"
 #include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/IsA.hpp"
 
 // IWYU pragma: no_forward_declare MakeWithValueImpl
 // IWYU pragma: no_forward_declare Variables
