+++ src/Domain/InterfaceComputeTags.hpp
@@ -22,6 +22,7 @@
 #include "Time/Tags.hpp"
 #include "Utilities/Gsl.hpp"
 #include "Utilities/TMPL.hpp"
+#include "Utilities/TypeTraits/IsA.hpp"
 
 namespace domain {
 namespace Tags {
