+++ src/NumericalAlgorithms/LinearOperators/Divergence.hpp
@@ -17,6 +17,7 @@
 #include "Utilities/Requires.hpp"
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TypeTraits.hpp"  // IWYU pragma: keep
+#include "Utilities/TypeTraits/IsA.hpp"
 
 /// \cond
 class DataVector;
