+++ src/NumericalAlgorithms/LinearOperators/PartialDerivatives.hpp
@@ -17,7 +17,7 @@
 #include "DataStructures/Variables.hpp"
 #include "Utilities/Requires.hpp"
 #include "Utilities/TMPL.hpp"
-#include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/IsA.hpp"
 
 /// \cond
 class DataVector;
