+++ src/Options/OptionsDetails.hpp
@@ -25,6 +25,7 @@
 #include "Utilities/Requires.hpp"
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/IsA.hpp"
 
 /// Holds details of the implementation of Options
 namespace Options_detail {
