+++ src/Parallel/ConstGlobalCache.hpp
@@ -16,6 +16,7 @@
 #include "Utilities/Requires.hpp"
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TaggedTuple.hpp"
+#include "Utilities/TypeTraits/IsA.hpp"
 
 #include "Parallel/ConstGlobalCache.decl.h"
