+++ src/Time/Actions/SelfStartActions.hpp
@@ -24,7 +24,7 @@
 #include "Utilities/Gsl.hpp"
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TaggedTuple.hpp"
-#include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/IsA.hpp"
 
 // IWYU pragma: no_include "DataStructures/Tensor/Tensor.hpp"
