+++ src/Utilities/Array.hpp
@@ -11,7 +11,7 @@
 
 #include "Utilities/Requires.hpp"
 #include "Utilities/TMPL.hpp"
-#include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/IsA.hpp"
 
 namespace cpp17 {
 namespace detail {
