+++ tests/Unit/Pypp/PyppFundamentals.hpp
@@ -34,6 +34,7 @@
 #include "Utilities/MakeArray.hpp"
 #include "Utilities/Requires.hpp"
 #include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/IsA.hpp"
 
 namespace pypp {
 /// \cond
