+++ tests/Unit/TestingFramework.hpp
@@ -19,7 +19,7 @@
 #include "Parallel/Abort.hpp"
 #include "Parallel/Exit.hpp"
 #include "Utilities/Requires.hpp"
-#include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/IsA.hpp"
 #include "Utilities/TypeTraits/IsIterable.hpp"
 #include "Utilities/TypeTraits/IsMaplike.hpp"
